Move and rename exclude-newer override type
