+++ crates/uv-distribution-types/src/lib.rs
@@ -63,6 +63,7 @@ pub use crate::dependency_metadata::*;
 pub use crate::diagnostic::*;
 pub use crate::dist_error::*;
 pub use crate::error::*;
+pub use crate::exclude_newer::*;
 pub use crate::file::*;
 pub use crate::hash::*;
 pub use crate::id::*;
@@ -94,6 +95,7 @@ mod dependency_metadata;
 mod diagnostic;
 mod dist_error;
 mod error;
+mod exclude_newer;
 mod file;
 mod hash;
 mod id;
