+++ crates/uv-resolver/src/lib.rs
@@ -1,10 +1,9 @@
 pub use dependency_mode::DependencyMode;
 pub use error::{ErrorTree, NoSolutionError, NoSolutionHeader, ResolveError, SentinelRange};
 pub use exclude_newer::{
-    ExcludeNewer, ExcludeNewerChange, ExcludeNewerPackage, ExcludeNewerPackageChange,
-    ExcludeNewerPackageEntry, ExcludeNewerSpan, ExcludeNewerValue, ExcludeNewerValueChange,
-    ExcludeNewerValueWithSpanRef, PackageExcludeNewer, PackageExcludeNewerChange,
-    serialize_exclude_newer_package_with_spans,
+    ExcludeNewer, ExcludeNewerChange, ExcludeNewerOverrideChange, ExcludeNewerPackage,
+    ExcludeNewerPackageChange, ExcludeNewerPackageEntry, ExcludeNewerValueChange,
+    ExcludeNewerValueWithSpanRef, serialize_exclude_newer_package_with_spans,
 };
 pub use exclusions::Exclusions;
 pub use flat_index::{FlatDistributions, FlatIndex};
@@ -30,6 +29,7 @@ pub use resolver::{
 };
 pub use universal_marker::{ConflictMarker, UniversalMarker};
 pub use upgrade::UpgradePackages;
+pub use uv_distribution_types::{ExcludeNewerOverride, ExcludeNewerSpan, ExcludeNewerValue};
 pub use version_map::VersionMap;
 pub use yanks::AllowedYanks;
